Add basic support for non-standard experimental pdu types
Instead of throwing an exception when non-standard pdu's are received, we
now parse the header and expose the body as a byte array. We can't do
much more than that.

This was noticed when receiving dis from JCATS 14.

JCATS sends a number of non-standard PDu's, one of them being PDU Type 137
which is an Org State pdu.