Move PduContainer and PduStream into disutil
Moving these as an ongoing effort to identify, analyze and potentially
deprecate some of the utility classes that have grown and to a degree
added noise and confusion for new users of the library.

Moving these two classes is potentially a breaking API change, but I think
the risk is low because the utility of these classes seems minimal and so
the likelihood anyone using these two classes is likely small.